refactor(core): extract Blockchain fork-activation mutations to own module
Two impl Blockchain methods move out of blockchain.rs to a new sibling
module crates/sentrix-core/src/blockchain_activations.rs:

  activate_evm      — one-shot AccountDB EVM migration at VOYAGER_EVM_HEIGHT
  activate_voyager  — Pioneer -> DPoS validator migration + epoch init

Both are guarded by persistent flags on the Blockchain struct
(evm_activated / voyager_activated) so post-fork restarts are no-ops.

Same multi-file impl Blockchain pattern as #659 / #660 / #661. Call
sites in the sentrix bin's validator loop keep working unchanged.

blockchain.rs drops to 2553 LOC (from 2619 post-#661, -66). 231 / 231
sentrix-core tests still green.
