+++ crates/sentrix-core/src/blockchain.rs
@@ -4,7 +4,6 @@ use crate::authority::AuthorityManager;
 use crate::vm::ContractRegistry;
 use sentrix_primitives::account::AccountDB;
 use sentrix_primitives::block::Block;
-use sentrix_primitives::error::SentrixResult;
 use sentrix_primitives::merkle::merkle_root;
 use sentrix_primitives::transaction::Transaction;
 use sentrix_storage::MdbxStorage;
@@ -485,72 +484,6 @@ impl Blockchain {
     /// Is the current chain past the EVM fork?
     pub fn is_evm_active(&self) -> bool {
         Self::is_evm_height(self.height())
-    }
-
-    /// Initialize EVM state at fork activation.
-    /// Called once when chain reaches VOYAGER_EVM_HEIGHT.
-    /// Migrates all account code_hash fields and initializes gas tracking.
-    /// Idempotent — guarded by the persistent `evm_activated` flag.
-    pub fn activate_evm(&mut self) {
-        if self.evm_activated {
-            tracing::debug!("activate_evm: already activated, skipping");
-            return;
-        }
-        tracing::info!("Activating EVM at height {}", self.height());
-        let migrated = self.accounts.migrate_to_evm();
-        self.evm_activated = true;
-        tracing::info!(
-            "EVM activated: {} accounts migrated, gas metering enabled",
-            migrated
-        );
-    }
-
-    /// Initialize Voyager state at fork activation.
-    /// Called once when chain reaches VOYAGER_DPOS_HEIGHT.
-    /// Migrates existing Pioneer validators to DPoS with equal stake.
-    /// Idempotent — guarded by the persistent `voyager_activated` flag so
-    /// validator restarts post-fork don't re-register validators or
-    /// re-snapshot the epoch.
-    pub fn activate_voyager(&mut self) -> SentrixResult<()> {
-        use sentrix_staking::MIN_SELF_STAKE;
-
-        if self.voyager_activated {
-            tracing::debug!("activate_voyager: already activated, skipping");
-            return Ok(());
-        }
-
-        // Migrate Pioneer validators → DPoS validators
-        let validators: Vec<String> = self
-            .authority
-            .active_validators()
-            .iter()
-            .map(|v| v.address.clone())
-            .collect();
-        for address in &validators {
-            if let Err(e) = self.stake_registry.register_validator(
-                address,
-                MIN_SELF_STAKE,
-                1000, // 10% default commission
-                self.height(),
-            ) {
-                tracing::warn!("Failed to migrate validator {}: {}", address, e);
-            }
-        }
-
-        // Initialize epoch manager with the new stake registry
-        self.stake_registry.update_active_set();
-        self.epoch_manager
-            .initialize(&self.stake_registry, self.height());
-
-        self.voyager_activated = true;
-
-        tracing::info!(
-            "Voyager DPoS activated at height {}. {} validators migrated.",
-            self.height(),
-            self.stake_registry.active_count()
-        );
-
-        Ok(())
     }
 
     // ── Chain validation ─────────────────────────────────
